# backend/src/main/java/fr/epita/assistants/ping/api/request/NewProjectRequest.java
package fr.epita.assistants.ping.api.request;

import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.AllArgsConstructor;

import java.util.List;
import java.util.UUID;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class NewProjectRequest {
    private String name;
}