# backend/src/main/java/fr/epita/assistants/ping/api/request/UserProjectRequest.java
package fr.epita.assistants.ping.api.request;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserProjectRequest {
    private UUID userId;
}