# backend/src/main/java/fr/epita/assistants/ping/api/response/ProjectResponse.java
package fr.epita.assistants.ping.api.response;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.List;
import java.util.UUID;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ProjectResponse {
    private UUID id;
    private String name;
    private List<UserSummaryResponse> members;
    private UserSummaryResponse owner;
}